Add files via upload
IoT and DevOps are the latest additions to the Information technology sector. DevOps is the combination of ‘Development and Operations’ while IoT represents the ‘Internet of Things’. DevOps is a complete software delivery procedure that emphasizes the collaboration between operations and the developers’ team. It includes product lifecycle management, developing operations, and software professionals.
DevOps specialize in automating the method of testing the products, integrating software, infrastructure, and deployment changes. the essential purpose is to develop a healthy environment where the creation, testing, and release of software can happen quickly and frequently. IoT talks about all the items connected to the web to decrease human intervention and energy.

The Connection Between IoT and DevOps
Read  More
https://blog.codingbrains.com/2020/11/24/devops-iotinternet-of-things-and-how-they-are-transforming-business/